requirements.txt
pyautogui
opencv-python
pillow
pathfinding
pytesseract